background

+++ balloon.py
@@ -45,7 +45,7 @@ class Balloon:
     
 
 def generate_random_balloon() -> Balloon:
-     return Balloon((random.randint(WIDTH*0.1, WIDTH*0.9), random.randint(HEIGHT, HEIGHT*1.2)),
+     return Balloon((random.randint(int(WIDTH*0.1), int(WIDTH*0.9)), random.randint(HEIGHT, int(HEIGHT*1.2))),
                     (random.randint(-20, 20), random.randint(-160, -80)), 
                      random.randint(20, 50), 
                     (random.randint(0, 255), random.randint(0, 255), random.randint(0, 255)))
